example/MODL/requirements.txt
absl-py==1.4.0
addict==2.4.0
altgraph==0.17
Augmentor==0.2.2
backports.entry-points-selectable==1.1.0
base==1.0.4
cachetools==5.3.0
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==3.0.1
cloudpickle==0.5.3
configparser==5.1.0
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.1
dask==0.18.1
distlib==0.3.3
docker-pycreds==0.4.0
dominate==2.6.0
einops==0.6.1
filelock==3.9.0
fonttools==4.39.3
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h5py==3.8.0
hdf5storage==0.1.19
huggingface-hub==0.12.1
idna==3.4
imageio==2.29.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.4
jsonpatch==1.32
jsonpointer==2.2
jupyterthemes==0.20.0
kiwisolver==1.4.4
lazy_loader==0.2
lesscpy==0.13.0
Markdown==3.4.1
MarkupSafe==2.1.2
mat73==0.35
matplotlib==3.7.1
MedPy==0.4.0
millify==0.1.1
ml-collections==0.1.1
mmcv-full==1.7.0
networkx==3.1
numpy==1.23.5
oauthlib==3.1.1
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
promise==2.3
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydevd-pycharm==222.3739.56
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.0
rsa==4.9
runstats==2.0.0
scikit-image==0.20.0
scipy==1.9.1
sentry-sdk==1.5.0
shortuuid==1.0.8
SimpleITK==2.1.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tifffile==2023.4.12
timm==0.6.12
toolz==0.9.0
torch==1.8.0+cu111
torchaudio==0.8.0
torchfile==0.1.0
torchvision==0.9.0+cu111
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.10.0
visdom==0.1.8.9
wandb==0.12.7
Werkzeug==2.2.3
yapf==0.32.0
yaspin==2.1.0
zipp==3.14.0
